+++ package.json
@@ -43,6 +43,7 @@
   "dependencies": {
     "@emotion/react": "^11.4.0",
     "@emotion/styled": "^11.3.0",
+    "@reduxjs/toolkit": "^1.6.0",
     "add": "^2.0.6",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
