+++ src/reducer.test.js
@@ -1,6 +1,4 @@
-import reducer from './reducer';
-
-import {
+import reducer, {
   setRegions,
   setCategories,
   setRestaurants,
@@ -13,7 +11,7 @@ import {
   changeReviewField,
   clearReviewFields,
   setReviews,
-} from './actions';
+} from './slice';
 
 describe('reducer', () => {
   context('when previous state is undefined', () => {
@@ -53,7 +51,6 @@ describe('reducer', () => {
       ];
 
       const state = reducer(initialState, setRegions(regions));
-
       expect(state.regions).toHaveLength(1);
     });
   });
